Fix IntegrationStepsHorizontalView and move details pages to new dir
Add flow prop
