+++ app/ui-react/syndesis/src/modules/integrations/pages/detail/index.ts
@@ -1,1 +1,3 @@
+export * from './ActivityPage';
 export * from './DetailsPage';
+export * from './MetricsPage';
