--- app/ui-react/syndesis/src/modules/integrations/pages/integration/detail/index.ts
@@ -1,3 +1,0 @@
-export * from './ActivityPage';
-export * from './DetailsPage';
-export * from './MetricsPage';
